decklink: Fix compiler warnings
CMake build framework 3.0 introduces more strict typing checks, which
require additional code fixes to silence.

* Fixes warnings about shadowing existing variables
* Fixes warnings about implicit casts
* Fixes warnings about missing newlines at the end of files

+++ plugins/decklink/decklink-device-discovery.cpp
@@ -94,7 +94,7 @@ DeckLinkDeviceDiscovery::DeckLinkDeviceRemoved(IDeckLink *device)
 
 ULONG STDMETHODCALLTYPE DeckLinkDeviceDiscovery::AddRef(void)
 {
-	return os_atomic_inc_long(&refCount);
+	return (ULONG)os_atomic_inc_long(&refCount);
 }
 
 HRESULT STDMETHODCALLTYPE DeckLinkDeviceDiscovery::QueryInterface(REFIID iid,
@@ -127,5 +127,5 @@ ULONG STDMETHODCALLTYPE DeckLinkDeviceDiscovery::Release(void)
 		return 0;
 	}
 
-	return newRefCount;
+	return (ULONG)newRefCount;
 }
